rm core folder, colocate inference types by concern
