+++ packages/core/src/http/types.ts
@@ -1,7 +1,10 @@
-import type { inferRouterContext, inferRouterError } from '../core/inference';
-import type { ProcedureType } from '../core/procedure';
-import type { AnyRouter } from '../core/router';
 import type { TRPCError } from '../error/TRPCError';
+import type { ProcedureType } from '../procedure';
+import type {
+  AnyRouter,
+  inferRouterContext,
+  inferRouterError,
+} from '../router';
 import type { TRPCResponse } from '../rpc';
 import type { Dict } from '../types';
