+++ packages/core/src/parser.ts
