+++ packages/core/src/procedure.ts
@@ -1,4 +1,8 @@
+import type { inferObservableValue } from './observable';
 import type { ProcedureCallOptions } from './procedureBuilder';
+import type { AnyRootConfig } from './rootConfig';
+import type { Serialize } from './serialize';
+import type { DefaultDataTransformer } from './transformer';
 
 export const procedureTypes = ['query', 'mutation', 'subscription'] as const;
 /**
@@ -71,3 +75,34 @@ export type AnyQueryProcedure = QueryProcedure<any>;
 export type AnyMutationProcedure = MutationProcedure<any>;
 export type AnySubscriptionProcedure = SubscriptionProcedure<any>;
 export type AnyProcedure = Procedure<ProcedureType, any>;
+
+/**
+ * @internal
+ */
+export type inferTransformedProcedureOutput<
+  TConfig extends AnyRootConfig,
+  TProcedure extends AnyProcedure,
+> = TConfig['transformer'] extends DefaultDataTransformer
+  ? Serialize<TProcedure['_def']['_output_out']>
+  : TProcedure['_def']['_output_out'];
+
+/** @internal */
+export type inferTransformedSubscriptionOutput<
+  TConfig extends AnyRootConfig,
+  TProcedure extends AnyProcedure,
+> = TConfig['transformer'] extends DefaultDataTransformer
+  ? Serialize<inferObservableValue<TProcedure['_def']['_output_out']>>
+  : inferObservableValue<TProcedure['_def']['_output_out']>;
+
+export type inferHandlerInput<TProcedure extends AnyProcedure> = ProcedureArgs<
+  inferProcedureParams<TProcedure>
+>;
+
+export type inferProcedureInput<TProcedure extends AnyProcedure> =
+  inferProcedureParams<TProcedure>['_input_in'];
+
+export type inferProcedureParams<TProcedure> = TProcedure extends AnyProcedure
+  ? TProcedure['_def']
+  : never;
+export type inferProcedureOutput<TProcedure> =
+  inferProcedureParams<TProcedure>['_output_out'];
